fix(tests): follow any-status host-canonicalization hops in SEO redirect specs
The seo-normalization spec's fetchCanonical only skipped the Vercel
apex→www hop when it arrived as a 307/302, single hop. In CI the hop is
a 308, so the spec asserted on the infrastructure redirect and all 15
cases failed with the source path echoed back in Location.

Extract the battle-tested multi-hop helper from redirects.spec.ts into
tests/helpers/fetch-canonical.ts (follows 301/302/307/308 same-path
cross-origin hops, max 3) and use it from both specs so the two can't
drift again. Verified: both specs pass against the apex domain, which
exercises the 308 hop that broke CI.
